requirements.txt
alembic==1.14.0
blinker==1.9.0
chardet==5.2.0
click==8.1.7
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
reportlab==4.2.5
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
